fix droppath error on training
